Updates to Trove loaders, makers, and loggers.

+++ loadTroveUsers.js
@@ -4,13 +4,19 @@ const readline = require('readline');
 const path = require('path');
 const { getAuthToken } = require('./lib/login');
 
-
 let inDir = process.argv[2];
 
-const wait = (ms) => {
-  console.log(`(Waiting ${ms}ms...)`);
-  return new Promise((resolve) => setTimeout(resolve, ms));
-};
+const cl = console.log;
+const logSeen = {}
+let logPath;
+console.log = (msg, path) => {
+  if (path) {
+    if (!logSeen[path] && fs.existsSync(path)) fs.unlinkSync(path);
+    fs.writeFileSync(path, msg + '\n', { flag: 'a'});
+    logSeen[path] = 1;
+  }
+  cl(msg);
+}
 
 (async () => {
   try {
@@ -30,6 +36,7 @@ const wait = (ms) => {
       }
       
       let config = await getAuthToken(superagent);
+      let ten = config.tenant;
 
       const fileStream = fs.createReadStream(inFile);
 
@@ -39,12 +46,12 @@ const wait = (ms) => {
       });
       
       let x = 0;
-      console.log('------------------------');
+      console.log('------------------------', logPath);
       for await (const line of rl) {
         x++;
         let rec = JSON.parse(line);
         let actionUrl = `${config.okapi}/${ep}`;
-        console.log(`[${x}] POST ${rec.id} to ${actionUrl}`);
+        console.log(`[${x}][${ten}] POST ${rec.id} to ${actionUrl}`, logPath);
         try {
           let res = await superagent
             .post(actionUrl)
@@ -52,11 +59,11 @@ const wait = (ms) => {
             .set('x-okapi-token', config.token)
             .set('content-type', 'application/json')
             .set('accept', 'application/json');
-          console.info(`  Successfully added record id ${rec.id}`);
+          console.log(`  Successfully added record id ${rec.id}`, logPath);
           success++;
         } catch (e) {
             let errMsg = (e.response && e.response.text) ? e.response.text : e;
-            console.log(errMsg);
+            console.log(errMsg, logPath);
             rec._errMessage = errMsg;
             let recStr = JSON.stringify(rec);
             fs.writeFileSync(errPath, recStr + '\n', { flag: 'a'});
@@ -69,8 +76,13 @@ const wait = (ms) => {
     let fail = 0;
 
     inDir = inDir.replace(/\/$/, '');
+    logPath = inDir + '/log.log';
     const cfile = inDir + '/config.json';
-    fs.copyFileSync(cfile, './config.json');
+    if (fs.existsSync(cfile)) {
+      fs.copyFileSync(cfile, './config.json');
+    } else {
+      throw new Error(`Can't find config file at ${inDir}!`);
+    }
 
     const ufile = inDir + '/users.jsonl';
     await postObjects(ufile, 'users');
@@ -78,7 +90,15 @@ const wait = (ms) => {
     const pfile = inDir + '/perms.jsonl';
     await postObjects(pfile, 'perms/users');
 
-    console.log('Success', success, 'Fail', fail);
+    console.log('----------------------------------', logPath);
+    console.log(`Success: ${success}`, logPath);
+    console.log(`Fail: ${fail}`, logPath);
+    console.log('----------------------------------', logPath);
+
+    let doneDir = inDir + '.DONE';
+    if (success > 0) {
+      fs.renameSync(inDir, doneDir);
+    }
 
   } catch (e) {
     console.log(`${e}`);
